implement game round event
